refactor(match): split the keyword tests to a sibling file for r8
The stopword work pushed `documents/keywords.rs` to 2303 LOC, past R8's 1400
hard cap, so `cargo test --test architecture` failed. Neither the `--lib` run
nor my own gate caught it, because both of us only ran `cargo test --lib`.

Pure move, no behaviour change: the `#[cfg(test)] mod test` block becomes
`documents/keywords/test.rs`, following the `commands/match_resume.rs` +
`match_resume/` precedent already in the tree. Parent drops to 1155 LOC, the
sibling is 1154, and the 42 keyword tests are unchanged and still pass.
